Exporting/Importing User Profile + Changing Album Views depending on loggedIn User
To import the user profile go to profile/readprofile.php page and click
the export button. The XML file will be saved to your XML folder. To
import the profile press the import button and upload your file. All the
user data will be deleted from the db and replaces with the XML records